refactor: split wandb sdk and publisher lifecycle
