+++ tools/measurement/measurement_tools/wandb_setup.py
@@ -1,40 +1,39 @@
 # SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
-"""Secure native W&B publishing for Anonymizer benchmark tooling."""
+"""Compatibility facade for secure native W&B publication."""
 
 from __future__ import annotations
 
 import logging
 import os
 import stat as stat
-import sys
-import threading
 from pathlib import Path
 from typing import Any, Callable
-from urllib.parse import quote
 
-from measurement_tools.wandb_models import (
-    DEFAULT_WANDB_PROJECT,
-    PUBLICATION_COMPLETE_KEY,
-    PUBLICATION_SEAL_DIGEST_KEY,
-    ResolvedWandbConfig,
-    WandbInitPayload,
-    WandbInputs,
-    WandbMode,
-    WandbPublicationState,
-    WandbPublishPayload,
-    WandbPublishResult,
-    WandbRunMetadata,
-)
-from measurement_tools.wandb_models import (
-    BenchmarkMetadata as BenchmarkMetadata,
-)
-from measurement_tools.wandb_models import (
-    WandbConfigPayload as WandbConfigPayload,
-)
+from measurement_tools import wandb_sdk_environment as _sdk_environment
+from measurement_tools.wandb_models import DEFAULT_WANDB_PROJECT as DEFAULT_WANDB_PROJECT
+from measurement_tools.wandb_models import BenchmarkMetadata as BenchmarkMetadata
+from measurement_tools.wandb_models import ResolvedWandbConfig as ResolvedWandbConfig
+from measurement_tools.wandb_models import WandbConfigPayload as WandbConfigPayload
+from measurement_tools.wandb_models import WandbInitPayload as WandbInitPayload
+from measurement_tools.wandb_models import WandbInputs as WandbInputs
+from measurement_tools.wandb_models import WandbMode as WandbMode
+from measurement_tools.wandb_models import WandbPublishPayload as WandbPublishPayload
+from measurement_tools.wandb_models import WandbPublishResult as WandbPublishResult
+from measurement_tools.wandb_models import WandbRunMetadata as WandbRunMetadata
 from measurement_tools.wandb_payload import BenchmarkWandbFinalization as BenchmarkWandbFinalization
 from measurement_tools.wandb_payload import build_publish_payload
+from measurement_tools.wandb_publisher import WandbPublisher as WandbPublisher
+from measurement_tools.wandb_publisher import (
+    define_benchmark_metrics,
+    publication_already_complete,
+    publication_state,
+    raise_lifecycle_failures,
+    sdk_init_kwargs,
+    wandb_run_url,
+)
 from measurement_tools.wandb_run_identity import default_run_name, effective_wandb_tags
+from measurement_tools.wandb_sdk_environment import WandbSdkEnvironment as WandbSdkEnvironment
 from measurement_tools.wandb_staging import open_directory_no_follow, validate_directory_metadata
 from measurement_tools.wandb_staging import prepare_wandb_staging_dir as _prepare_wandb_staging_dir
 
@@ -48,232 +47,34 @@ __all__ = [
 
 logger = logging.getLogger("measurement.wandb")
 
+WANDB_SANITIZER_VERSION = 2
+_WANDB_INSTALL_HINT = _sdk_environment._WANDB_INSTALL_HINT
+_PUBLISHER_WANDB_MODULE: Any | None = _sdk_environment._PUBLISHER_WANDB_MODULE
+_WANDB_AMBIENT_ALLOWLIST = _sdk_environment._WANDB_AMBIENT_ALLOWLIST
+_PROCESS_ENVIRONMENT_ALLOWLIST = _sdk_environment._PROCESS_ENVIRONMENT_ALLOWLIST
+_WANDB_ENVIRONMENT_LOCK = _sdk_environment._WANDB_ENVIRONMENT_LOCK
+_WANDB_ENVIRONMENT_OWNER = _sdk_environment._WANDB_ENVIRONMENT_OWNER
+
 _default_run_name = default_run_name
 _effective_wandb_tags = effective_wandb_tags
-
-WANDB_SANITIZER_VERSION = 2
-_WANDB_INSTALL_HINT = "Install the optional measurement dependency group: uv sync --group measurement"
-
-
-_PUBLISHER_WANDB_MODULE: Any | None = None
-_RESUMABLE_CONFIG_REFRESH_KEYS = frozenset(
-    {
-        "benchmark_identity",
-        "benchmark_role",
-        "benchmark_kind",
-        "suite_version",
-        "branch",
-        "commit_sha",
-        "commit_short",
-        "pr_number",
-        "release_version",
-        "anonymizer_config_id",
-        "anonymizer_mode",
-    }
-)
+_publisher_environment = _sdk_environment.publisher_environment
+_publication_already_complete = publication_already_complete
+_publication_state = publication_state
+_wandb_run_url = wandb_run_url
+_raise_lifecycle_failures = raise_lifecycle_failures
+_sdk_init_kwargs = sdk_init_kwargs
+_define_benchmark_metrics = define_benchmark_metrics
 
 
 def require_wandb() -> Any:
-    """Import wandb inside the publisher's guarded SDK environment."""
+    """Use the canonical SDK loader while preserving the legacy patch sentinel."""
     global _PUBLISHER_WANDB_MODULE  # noqa: PLW0603
 
-    if _WANDB_ENVIRONMENT_OWNER != threading.get_ident():
-        raise RuntimeError("wandb must be imported inside the guarded SDK environment")
-    preloaded = sys.modules.get("wandb")
-    if _PUBLISHER_WANDB_MODULE is None and preloaded is not None:
-        raise RuntimeError("wandb must not be imported before the guarded publisher")
-    if _PUBLISHER_WANDB_MODULE is not None and preloaded is not _PUBLISHER_WANDB_MODULE:
-        raise RuntimeError("the loaded wandb module changed outside the guarded publisher")
+    _sdk_environment._PUBLISHER_WANDB_MODULE = _PUBLISHER_WANDB_MODULE
     try:
-        import wandb
-    except ImportError as exc:
-        raise ImportError(
-            f"W&B logging is enabled but the wandb package is not installed. {_WANDB_INSTALL_HINT}"
-        ) from exc
-    _PUBLISHER_WANDB_MODULE = wandb
-    return wandb
-
-
-_WANDB_AMBIENT_ALLOWLIST = frozenset(
-    {
-        "WANDB_API_KEY",
-        "WANDB_HTTP_TIMEOUT",
-        "WANDB_INIT_TIMEOUT",
-        "WANDB__SERVICE_WAIT",
-    }
-)
-_PROCESS_ENVIRONMENT_ALLOWLIST = frozenset(
-    {
-        "HOME",
-        "LANG",
-        "LC_ALL",
-        "LC_CTYPE",
-        "PATH",
-        "TEMP",
-        "TMP",
-        "TMPDIR",
-        "TZ",
-    }
-)
-_WANDB_ENVIRONMENT_LOCK = threading.Lock()
-_WANDB_ENVIRONMENT_OWNER: int | None = None
-
-
-class WandbSdkEnvironment:
-    """Process-wide exact environment transaction around W&B SDK use."""
-
-    def __init__(self, settings: ResolvedWandbConfig) -> None:
-        self._settings = settings
-        self._snapshot: dict[str, str] | None = None
-
-    def __enter__(self) -> WandbSdkEnvironment:
-        global _WANDB_ENVIRONMENT_OWNER  # noqa: PLW0603
-
-        if not _WANDB_ENVIRONMENT_LOCK.acquire(blocking=False):
-            raise RuntimeError("nested or concurrent W&B publisher use is not allowed")
-        _WANDB_ENVIRONMENT_OWNER = threading.get_ident()
-        self._snapshot = dict(os.environ)
-        try:
-            preserved = {
-                key: value
-                for key, value in self._snapshot.items()
-                if key in _PROCESS_ENVIRONMENT_ALLOWLIST or key in _WANDB_AMBIENT_ALLOWLIST
-            }
-            os.environ.clear()
-            os.environ.update(preserved)
-            os.environ.update(_publisher_environment(self._settings))
-        except BaseException:
-            self._restore()
-            raise
-        return self
-
-    def __exit__(self, _exc_type: Any, _exc: Any, _traceback: Any) -> None:
-        self._restore()
-
-    def _restore(self) -> None:
-        global _WANDB_ENVIRONMENT_OWNER  # noqa: PLW0603
-
-        if self._snapshot is None:
-            return
-        os.environ.clear()
-        os.environ.update(self._snapshot)
-        self._snapshot = None
-        _WANDB_ENVIRONMENT_OWNER = None
-        _WANDB_ENVIRONMENT_LOCK.release()
-
-
-def _publisher_environment(settings: ResolvedWandbConfig) -> dict[str, str]:
-    environment = {
-        "WANDB_ERROR_REPORTING": "false",
-        "WANDB_MODE": settings.wandb_mode.value,
-        "WANDB_PROJECT": settings.wandb_project,
-        "WANDB_SILENT": "true",
-    }
-    optional = {
-        "WANDB_BASE_URL": settings.wandb_base_url,
-        "WANDB_ENTITY": settings.wandb_entity,
-        "WANDB_GROUP": settings.wandb_group,
-        "WANDB_JOB_TYPE": settings.wandb_job_type,
-        "WANDB_NAME": settings.wandb_run_name,
-        "WANDB_TAGS": settings.wandb_tags or None,
-    }
-    environment.update({key: value for key, value in optional.items() if value is not None})
-    return environment
-
-
-class WandbPublisher:
-    """Own the complete strict native W&B publication lifecycle."""
-
-    def publish(
-        self,
-        settings: ResolvedWandbConfig,
-        *,
-        suite_id: str,
-        output_dir: Path,
-        finalization: BenchmarkWandbFinalization,
-        metadata: WandbRunMetadata | None = None,
-    ) -> WandbPublishResult:
-        if not settings.enabled:
-            return WandbPublishResult(published=False)
-        payload, snapshot_sha256, record_count = _build_publish_payload(
-            settings,
-            suite_id=suite_id,
-            output_dir=output_dir,
-            finalization=finalization,
-            metadata=metadata,
-        )
-        return self.publish_payload(
-            settings,
-            payload=payload,
-            measurement_sha256=snapshot_sha256,
-            record_count=record_count,
-        )
-
-    def publish_payload(
-        self,
-        settings: ResolvedWandbConfig,
-        *,
-        payload: WandbPublishPayload,
-        measurement_sha256: str,
-        record_count: int,
-    ) -> WandbPublishResult:
-        """Publish a complete typed payload without access to source artifacts."""
-        if not settings.enabled:
-            return WandbPublishResult(published=False)
-        with WandbSdkEnvironment(settings):
-            wandb = require_wandb()
-            if getattr(wandb, "run", None) is not None:
-                raise RuntimeError("an ambient W&B run is active")
-            run: Any | None = None
-            result: WandbPublishResult | None = None
-            primary_error: BaseException | None = None
-            try:
-                run = wandb.init(**_sdk_init_kwargs(wandb, payload.init))
-                if run is None:
-                    raise RuntimeError("wandb.init did not return an explicit run handle")
-                run_id = str(getattr(run, "id", ""))
-                if run_id != payload.init.run_id:
-                    raise RuntimeError("wandb.init returned a different run identity")
-                already_complete = _publication_already_complete(run, payload)
-                publication_state = _publication_state(run, already_complete=already_complete)
-                if already_complete:
-                    _refresh_resumed_complete_config(run, payload)
-                else:
-                    _define_benchmark_metrics(run)
-                    run.config.update(payload.config.sdk_values(), allow_val_change=True)
-                    tables = {
-                        table.name: wandb.Table(columns=list(table.columns), data=[list(row) for row in table.data])
-                        for table in payload.tables
-                    }
-                    logged = {**payload.history.metrics, **tables}
-                    if payload.init.resume == "allow":
-                        run.log(logged, step=0)
-                    else:
-                        run.log(logged)
-                    run.summary.update(payload.summary.metrics)
-                logger.info("W&B run id: %s", run_id)
-                result = WandbPublishResult(
-                    published=True,
-                    run_id=run_id,
-                    entity=settings.wandb_entity,
-                    project=settings.wandb_project,
-                    run_url=_wandb_run_url(settings, run_id=run_id),
-                    publication_state=publication_state,
-                    measurement_sha256=measurement_sha256,
-                    record_count=record_count,
-                )
-            except BaseException as exc:
-                primary_error = exc
-            finish_error: BaseException | None = None
-            if run is not None:
-                try:
-                    run.finish()
-                except BaseException as exc:
-                    finish_error = exc
-            _raise_lifecycle_failures(primary_error, finish_error)
-            if result is None:
-                raise RuntimeError("W&B publisher completed without a result")
-            return result
+        return _sdk_environment.require_wandb()
+    finally:
+        _PUBLISHER_WANDB_MODULE = _sdk_environment._PUBLISHER_WANDB_MODULE
 
 
 def publish_benchmark_wandb_best_effort(
@@ -321,107 +122,6 @@ def _build_publish_payload(
     )
 
 
-def _publication_already_complete(run: Any, payload: WandbPublishPayload) -> bool:
-    if payload.init.resume != "allow":
-        return False
-    expected_digest = payload.summary.metrics[PUBLICATION_SEAL_DIGEST_KEY]
-    summary = getattr(run, "summary", None)
-    get_value = getattr(summary, "get", None)
-    config = getattr(run, "config", None)
-    get_config = getattr(config, "get", None)
-    if not callable(get_value) or not callable(get_config):
-        raise RuntimeError("resumed W&B run does not expose readable publication state")
-    observed_complete = get_value(PUBLICATION_COMPLETE_KEY)
-    observed_digest = get_value(PUBLICATION_SEAL_DIGEST_KEY)
-    observed_imported = get_config("imported")
-    observed_config_digest = (
-        observed_imported.get("completion_seal_sha256") if isinstance(observed_imported, dict) else None
-    )
-    if observed_config_digest not in {None, expected_digest}:
-        raise RuntimeError("resumed W&B run config contains different sealed content")
-    if observed_complete is True:
-        if observed_digest != expected_digest or observed_config_digest != expected_digest:
-            raise RuntimeError("resumed W&B run is complete for different sealed content")
-        return True
-    if observed_complete not in {None, False}:
-        raise RuntimeError("resumed W&B run has an invalid publication marker")
-    if observed_digest not in {None, expected_digest}:
-        raise RuntimeError("resumed W&B run contains different sealed content")
-    return False
-
-
-def _refresh_resumed_complete_config(run: Any, payload: WandbPublishPayload) -> None:
-    desired = {
-        key: value for key, value in payload.config.sdk_values().items() if key in _RESUMABLE_CONFIG_REFRESH_KEYS
-    }
-    if not desired:
-        return
-    config = getattr(run, "config", None)
-    get_config = getattr(config, "get", None)
-    update_config = getattr(config, "update", None)
-    if not callable(get_config) or not callable(update_config):
-        raise RuntimeError("resumed W&B run does not expose mutable publication config")
-    delta = {key: value for key, value in desired.items() if get_config(key) != value}
-    if delta:
-        update_config(delta, allow_val_change=True)
-
-
-def _publication_state(run: Any, *, already_complete: bool) -> WandbPublicationState:
-    if already_complete:
-        return WandbPublicationState.already_complete
-    return WandbPublicationState.resumed if bool(getattr(run, "resumed", False)) else WandbPublicationState.created
-
-
-def _wandb_run_url(settings: ResolvedWandbConfig, *, run_id: str) -> str | None:
-    if settings.wandb_mode != WandbMode.online or settings.wandb_entity is None:
-        return None
-    base_url = settings.wandb_base_url or "https://wandb.ai"
-    if base_url == "https://api.wandb.ai":
-        base_url = "https://wandb.ai"
-    segments = (settings.wandb_entity, settings.wandb_project, "runs", run_id)
-    return f"{base_url}/{'/'.join(quote(segment, safe='') for segment in segments)}"
-
-
-def _raise_lifecycle_failures(primary: BaseException | None, finish: BaseException | None) -> None:
-    if primary is not None and finish is not None:
-        if isinstance(primary, Exception) and isinstance(finish, Exception):
-            raise ExceptionGroup("W&B publication and finish both failed", [primary, finish]) from primary
-        raise BaseExceptionGroup("W&B publication and finish both failed", [primary, finish]) from primary
-    if primary is not None:
-        raise primary.with_traceback(primary.__traceback__)
-    if finish is not None:
-        raise finish.with_traceback(finish.__traceback__)
-
-
-def _sdk_init_kwargs(wandb: Any, payload: WandbInitPayload) -> dict[str, Any]:
-    values: dict[str, Any] = {
-        "project": payload.project,
-        "id": payload.run_id,
-        "resume": payload.resume,
-        "name": payload.name,
-        "mode": payload.mode.value,
-        "settings": wandb.Settings(
-            console="off",
-            disable_code=True,
-            disable_git=True,
-            host="redacted",
-            save_code=False,
-            x_disable_machine_info=True,
-            x_disable_meta=True,
-            x_disable_stats=True,
-            x_save_requirements=False,
-        ),
-        "dir": str(payload.directory),
-        "group": payload.group,
-        "job_type": payload.job_type,
-    }
-    if payload.entity is not None:
-        values["entity"] = payload.entity
-    if payload.tags:
-        values["tags"] = list(payload.tags)
-    return values
-
-
 def prepare_wandb_staging_dir(output_dir: Path) -> Path:
     """Prepare secure W&B staging through the canonical filesystem owner."""
     return _prepare_wandb_staging_dir(output_dir)
@@ -435,14 +135,3 @@ def _open_directory_no_follow(path: Path) -> int:
 def _validate_directory_metadata(metadata: os.stat_result, *, final: bool) -> None:
     """Compatibility alias for canonical directory metadata validation."""
     validate_directory_metadata(metadata, final=final)
-
-
-def _define_benchmark_metrics(run: Any) -> None:
-    define_metric = getattr(run, "define_metric", None)
-    if not callable(define_metric):
-        return
-    for metric_name in ("benchmark/*", "measurement/*"):
-        try:
-            define_metric(metric_name, summary="last")
-        except Exception as exc:  # noqa: BLE001 -- presentation polish is best-effort
-            logger.warning("Failed to define W&B metric %s (%s)", metric_name, type(exc).__name__)
